requirements.txt
# Core scientific stack
numpy>=2.1.0,<3.0.0
pandas>=2.2.2
scipy>=1.15.0

# ML / NLP
torch>=2.0.0
transformers>=4.40.0,<5.0.0
scikit-learn>=1.6.0

# Hugging Face datasets
datasets>=4.0.0,<5.0.0

# LLM / API
openai>=1.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0

# RAG retrieval
rank-bm25>=0.2.2

# RAG evaluation
ragas>=0.3.0,<0.4.0

# Vector DB
chromadb>=1.0.0

# Text processing
tqdm>=4.60.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0
